requirements.txt
streamlit
pandas
numpy
openai
python-dotenv
scipy